requirements.txt
tornado>=4.0.0

# optional
pushbullet.py>=0.8.0
chump>=1.5.0
xmpppy
easygui
requests
python-telegram-bot
